+++ NEWS.md
@@ -3,3 +3,7 @@
 * Added a `NEWS.md` file to track changes to the package.
 * Added example data in a `xcmsRawLike` format
 * Added a Vignette showing the workflow based on the example data
+* `DeconvoluteSpectrum()` can handle now a missing `mz1` parameter
+* `getMultipleBPC` better handles and unspecified `mz` parameter
+* `EvaluatePairsFromXCMSSet()` now also accepts simple peak lists to be more
+  independent of `xcms`